Provide new default for runtime identifier Fixes #35575 (#36848)
Recreated #36149 to avoid pinging everyone as I changed the base from 8.0.2xx to main.

Fixes #35575

This block of code essentially set a default for the runtime identifier if none was specified. It did not take into account that, when users do not specifically request legacy rids, we shifted the default to use the simplified rid graph. This now does that.