Add filtering to SdkSupportedTargetPlatformVersion validation. Fixes #38016.
The SdkSupportedTargetPlatformVersion item group is used for (at least) two things:

1. Generate the _OR_GREATER preprocessing symbols:

https://github.com/dotnet/sdk/blob/bfd2919bc446cad68d11de90cec4025d3683591c/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Sdk.BeforeCommon.targets#L230-L237

2. Validate the TargetPlatformVersion:

https://github.com/dotnet/sdk/blob/bfd2919bc446cad68d11de90cec4025d3683591c/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.TargetFrameworkInference.targets#L233-L246

The problem is that these two uses aren't equivalent.

Take for example the following scenario:

We release bindings for iOS 10, and a library developer takes advantage of the
bindings for the new iOS version, while at the same time supporting
multi-targeting to older platforms:

```csharp
    #if IOS10_0_OR_GREATER
        UseNewApi ();
    #else
        UseOldApi ();
    #endif
```

Time passes, iOS 11 comes out, and we stop shipping bindings specifically for
iOS 10 (the APIs themselves would be included in the bindings for iOS 11). The
code above should continue to work, but iOS 10 is not a valid
TargetPlatformVersion anymore. However, with the current situation there's no
way to express this, because the moment we remove the "10.0" version from
SdkSupportedTargetPlatformVersion, the IOS10_0_OR_GREATER define isn't
generated anymore.

We discussed this in a meeting internally, and the suggestion that came up
was to use metadata to handle this situation.

So to solve this, I've added metadata to the SdkSupportedTargetPlatformVersion
item group to indicate that a particular value is only valid to produce the
_OR_GREATER processing symbol:

```xml
<ItemGroup>
    <SdkSupportedTargetPlatformVersion Include="10.0" OrGreaterHistoricalValue="true" />
    <SdkSupportedTargetPlatformVersion Include="12.0" />
</ItemGroup>
```

And then filter out these historical values when validating the
TargetPlatformVersion.

Fixes https://github.com/dotnet/sdk/issues/38016.