Support glob patterns when ignoring baseline TFMs
Change the task and targets input to a property instead of an item.
This feature hasn't yet shipped nor got documented so it doesn't require a
breaking change noticed.

The new property ' PackageValidationBaselineFrameworksToIgnore' now
supports the glob '*' character. Multiple entries should be separated
via the ';' character. This is documented in both frontends.

Added tests to verify the behavior. The comparison is still performed
with ignored casing and ignored culture.