Use System.Text.Json in Netfx build of MSBuildSdkResolver (#39573)
### Summary

Switch the .NET Framework version of `MSBuildSdkResolver` to use `System.Text.Json` instead of `Newtonsoft.Json` to align it with MSBuild and make it possible to load the resolver with `Assembly.Load`.

### Notes

We cannot build against the SDK version of S.T.J because that's generally not available in MSBuild. I have set the version to 8.0.0 and I believe it can stay like this for as long as we don't need any newer functionality from the library. When the need to upgrade arises, we should just check that MSBuild in VS where this is inserted to has the new (or higher) version bundled with it.