Clean up contents from previously-sb tarball post binary-cleanup
If we extracted the previously-source-build tarball only for running the
binary cleanup tool, clean up the extracted contents. This undo's the
extraction operation that was only performed for running the
binary-cleanup tool.

I am creating a tarball for building .NET offline post-prep. In that
scenario, I run ./prep-source-build.sh --bootstrap, and then package up
the entire VMR directory into a tarball. Without this cleanup, there are
a ton of duplicated binaries in the tarball, since the nuget packages
are both in previously-source-built archive and in an extracted form.
With this change, the archive size goes from just over 4GB to just over
2GB.

This is a port of https://github.com/dotnet/installer/pull/19632 from
installer repo to sdk.