Re-enable test that was blocking codeflow
This test is running `GetValuesCommand` with `PublishTrimmed` set
to `true` in the generated
`<project>.WriteValuesToFile.g.targets`. In the failure case,
`ImportProjectExtensionProps` is false (due to
https://github.com/dotnet/msbuild/blob/147ecadd19ae031d5a511ad55908cff9bcdc17c5/src/Tasks/Microsoft.Common.props#L68
imported from full framework MSBuild props) during restore, so
this file isn't imported, we don't set `PublishTrimmed`,
Microsoft.NET.ILLink.Tasks isn't restored, and then
`ILLinkTargetsPath` never gets set.

I was wondering why this wasn't failing in main on .NET Core, and
found that the `GetValuesCommand` behavior was fixed as part of
https://github.com/dotnet/sdk/pull/39088 to not rely on
`ImportProjectExtensionProps`. So an alternative fix would be to backport the GetValuesCommand fix.

This change is just a simpler fix to set `PublishTrimmed` in the
project file for this testcase.