Use the VMR commit hash for dotnet --info
Remove the logic that overrides the git commit info for each repo. Let
each repo infer the VMR's git commit (via sourcelink).

That will let the sdk, runtime, aspnetcore (and other) repos use the
VMR's commit hash and show that in dotnet --info.

This works for both when building the VMR directly in a git repository,
as well as when building it from a git-archive tarball thanks to
sourcelink's support for minimal git metadata [1].

This is a port of https://github.com/dotnet/installer/pull/18941 to
dotnet/sdk since the installer was merged into sdk.

[1] https://github.com/dotnet/sourcelink/tree/main/docs#minimal-git-repository-metadata

Fixes: dotnet/source-build#3643