Minimize netfx containers task dependencies
Some of these references are relevant only to the "library" version of
this assembly and not needed by the net472 version of the task, after
ifdefing some uncalled code from the task.

System.Text.Json is more interesting: it's provided by MSBuild, but at
a version that may be slightly stale, so reference the current such
version.