Add Android workload manifests to SB .NET SDK
Context: https://github.com/dotnet/source-build/issues/3242

If using a source-built .NET SDK, the Android workload manifests are
not included in the SDK. So, you won't see `android` workload when you
run `dotnet workload search`.

The Android workload, in general, contains private sources, but the
manifest files are completely public.

In the long run, we could consider onboarding dotnet/android to:

* https://github.com/dotnet/source-build/blob/main/Documentation/sourcebuild-in-repos/new-repo.md

But the prerequisites would be to first have the Android workload
onboard to arcade:

* https://github.com/dotnet/arcade/blob/main/Documentation/Onboarding.md

*For now*, we could include the Android workload in source-build by:

* Add a step that copies the Android workload manifests source code
  in-tree (to be used when updating manifests).

* Commit the manifest sources.

* Source-build will then include the Android workload manifests in
  the SDK by using the source code, in-tree.