No .deps.json for net472 Container task
MSBuild only respects a .deps.json in the net8.0 engine, so there's no
need to create it for a net472 task. It includes references that aren't
part of the SDK, like the downlevel `System.Text.Json` that is provided
by MSBuild/Visual Studio, so removing it can avoid false positives from
dependency scanners.