Do not go beyond the root when searching for a previously installed tool Fixes #40655 (#40800)
Fixes #40655

dotnet tool install currently looks for a previously installed tools manifest. It reads manifests it finds in looking for one that has the manifest it's trying to install but does not check whether that manifest 'isRoot', which means it keeps going outside the root.

We don't use a tool beyond the root, which means we can't install then use a tool as would be expected.