Clean up STJ references Fixes #39902 (#40645)
Change our System.Text.Json references to align with what is present in the SDK (core—which just means not doing anything) or VS (framework—and align with the minimum version in the VS we're shipped with). Task projects do not need the override. See https://github.com/dotnet/sdk/pull/40645#discussion_r1657825731