Define "ToolingPackageVersion" properties for msbuild tooling dependencies
This change adds a set of properties to define the version of the msbuild tooling dependencies.
This is done to make it easier to update the versions of these dependencies in the future
and to use a common set of versions across all projects.

Upgrade the hardcoded 15.7.179 msbuild dependencies
to 17.8.3 to avoid transitive netstandard1.x
dependencies. This makes the building sdk
form source netstandard1.x clean.

Also set FlagNetStandard1XDependencies to avoid
accidentally brining in netstandard1.x dependencies
in the source-build.