Simplify MSBuildSdkResolver
Don't create a nuget package for it which isn't used anymore in the
stack and was non-shipping to begin with.

Don't multi-target and remove the .NETCoreApp TFM which isn't used
anymore. Same in the test project.

Remove the project from the source-build slnf

Remove outdated comments inSetBuildDefaults that referred to the
project.

Let VSMSBuildExtensions consume the assets directly from
MSBuildSdkResolver instead of downloading them yet again which could
result in a disjoint.