Fix updating KnownWebAssemblySdkPack when creating test layout
We missed adding KnownWebAssemblySdkPack into OverrideAndCreateBundledNETCoreAppPackageVersion so it was pointing to an older version in the BundledVersion.props in the test layout.

Fix that and add some checks so we don't miss this again the next time we add a Known*Pack entry.

Fixes https://github.com/dotnet/sdk/issues/44886