Write APICompat suppressions when baseline suppressions aren't empty
When there aren't any suppressions but baseline suppressions exist,
write to the suppression file to make it empty.

Pass information back to the caller so that even if an empty collection
of suppression are written to the file, still log a message to indicate
that the suppression file got updated.