[FIX MAIN] Update the tests to reflect the new templates.
Resolves https://github.com/dotnet/sdk/issues/45283

Ok so the template engine uses 'Verify' which I dont see elsewhere in the SDK --
What this does is store a txt snapshot of the last test result.

It looks like Class, Record, Enum, Struct, etc may be new templates added in net 10 that need to be added to the test logic.
https://github.com/dotnet/sdk/pull/29655
They were actually added in 7.0.200.  I dont know why but the last test result didnt have the templates added in 7.0.200 for net 10, maybe they were not updated and it got temporarily removed. I couldnt find that in the history, though, so I've added them now that theyre showing up.

I dont think this test will pass because there is a whitespace output change as well. Im not sure if that applies to mac and linux or not.
`DotnetNewListTests.BasicTest_WhenListCommandIsUsed.OSX.verified.txt` will have to be updated and `DotnetNewListTests.BasicTest_WhenListCommandIsUsed.Linux.verified.txt` will have to be with the new templates, if those fail, otherwise if the legacy tests also fail, they will have to be updated with the new templates and new whitespace.