Fix BuildPass>1 artifacts and asset manifest content
Fixes https://github.com/dotnet/source-build/issues/4778

There are three problems with the current implementation:

1. The asset manifest that gets produced in a BuildPass>2 build contains
all the assets from the previous dependent vertical manifests. This is
undesirable in general but also causes an issue when we merge all manifests
together because there are duplicates that can't be resolved.

2. Artifacts from the dependent verticals get uploaded to the
"<JobName>_Artifacts" pipeline artifact in CI in addition to the new
artifacts produced which is a waste of resources. It also makes it harder
to differentiate what actually got produced in the BuildPass>1 leg.

3. The log that lists all the artifacts that got produced by the repo build
also outputs the artifacts from the dependent verticals because
their manifests were considered as well.

4. Repo asset manifests from dependent verticals have the same name and
therefore got ignored. Only the first one gets copied, all the other
repo asset manifest files are ignored.

---

Isolate the asset manifests that get produced by the build and then later
read to generated the merged (vertical) manifest, from the downloaded dependent
vertical manifests. Upload the vertical manifest only, instead of all the inner
repo manifests. Name the manifest based on the job name to avoid conflicts.

Don't upload everything under artifacts/assets and artifacts/packages. Instead
read from the current vertical merged manifest to identify the assets to upload.
Leverage an msbuild project for this that prepares the directory to upload. As part
of that, also introduce a YML configuration property that defaults to Release and
pass that into the build to be able to find the path to the vertical manifest without
hardcoding a certain configuration.