Don't require there to be a supported runtime pack runtime identifier for crossgen2
We already resolve the correct crossgen2 for the host in ProcessFrameworkReferences.

This resolves an issue that can be encountered in a VMR-based cross-build source-build. In this scenario, we have a crossgen2 that can run on the host and a runtime pack that can run on the target, but no runtime pack for the host machine. This is okay as long as we use crossgen2 (which we would always be using in this scenario). However, before this change, if we explicitly define the RIDs that we have packs for to match the exact packs that are available (which aspnetcore does), then the SDK would fail to run crossgen2.