Do not automatically add nuget.org as a package source when it is absent from nuget.config Fixes #44312 (#45410)
Fixes #44312

If a user doesn't have nuget.org as a source in their nuget.config, we shouldn't assume they want it. This sends a message for dotnet tool search if nuget.org is absent, as it is needed for dotnet tool search to work properly, instead of running the command. For other commands like dotnet new details, it does not automatically add nuget.org to the sources.

image

image

(Neither command implicitly went to nuget.org.)