Shorten the sdk layout path folder
We are hitting MAX_PATH limits in the VMR when building the sdk repository with an OfficialBuildId that has three digits.

Shortening the "internal" name to "int" was the best immediate solution. We will eventually switch to Wix v5/6 which will avoid these issues alltogether as it won't have MAX_PATH restrictions.