Migrate sdk's sign/publish infra to use the Artifacts extension point
With improvements in infrastructure, we can utilize the new Artifact extension point to handle checksum generation and blob placement without calling PushToBuildStorage manually.

Also, there were two separate places where the repo was persisting the non-shipping version number (from before the sdk and installer merge). Update the build to only persist it once and read it in an evaluation-time manner.