Add CA1873: Avoid potentially expensive logging (#7290)
* Add CA1873: Avoid potentially expensive logging

This analyzer detects calls to 'ILogger.Log', extension methods in
'Microsoft.Extensions.Logging.LoggerExtensions' and methods decorated
with '[LoggerMessage]'.
It then checks if they evaluate expensive arguments without checking if
logging is enabled with 'ILogger.IsEnabled'.

* Use loop for IsGuardedByIsEnabled instead of recursion

* Rename TryGetLogLevel to IsLogInvocation

* Rework how expensive argument evaluation is detected

Previously, some specific operations were treated as non-expensive and
the rest were. This is not really future-proof as the language evolves,
and is generally more prone to false positives.

So this has now been changed to the opposite: Treat certain operations
as expensive, and the rest are not expensive by default and do not
trigger a diagnostic.

* Fix same instance detection when using conditional access

* Run msbuild pack

* Reorder tests

* Fix constant interpolation string detection

* Cleanup and add missing tests

* Add StringSyntax attribute to highlight test source

* Allow nameof, consts and literals in interpolated strings

* Use SymbolEqualityComparer to check same instance symbol

* Check all parent blocks for IsEnabled guard