Add lock (dotnet/templating#8635)
Fixes dotnet/templating#7946

dsplaisted found a stack here that was very helpful in helping me find the source of the parallel call to the workload resolver. As seen in the top stack, this call fairly quickly reaches the WorkloadResolver, where it attempts to initialize everything concurrently, causing the issue. It's called from TemplatePackageCoordinator in the SDK repo, which in turn is called by various things.

This adds a lock around the call to the WorkloadResolver in WorkloadConstraintFactory, which should prevent it from trying to go through the initialization code in parallel.

Commit migrated from https://github.com/dotnet/templating/commit/8c5b547aae29264f91da1a90466b050167d7389e