Switch verb and noun for package/reference add/remove/list Fixes #9650 (#45384)
Switches the ordering of commands from dotnet verb noun to dotnet noun verb, as it is sometimes unclear what 'add' adds, for instance, and this is how many users of other CLIs think.

In this version, dotnet verb noun (DVN) is still supported in that it works when used, but dotnet noun verb (DNV) is the priority. DVN is now hidden and doesn't appear in the help text, whereas DNV is visible and does appear in the help text.

We will start attempting to move users to the new version in a later version.