Remove app host packages for different archs from packs folder
These apphost packages are now getting installed via the apphost msi and the bundle. Remove them from the pre-downloaded "packs" folder so that the SDK archive can be fully built from a single vertical. This was the only place that did require a join to construct the SDK archive itself.