Fix SDKBundleVersion calculation
Previously, redist-installer produced an AssemblyInfo.cs file. That isn't the case anymore as the project now uses the Microsoft.Build.NoTargets SDK as it doesn't need to compile anything. Because of that, the `FileVersion` property doesn't get set automatically anymore.

Define a dependency on the `GetAssemblyVersion` target in the SDK to get the `FileVersion` property set in time.