Change property used for after-publish target (#47068)
This is based on feedback:
https://github.com/dotnet/sdk/pull/36850#issuecomment-2676884372
https://github.com/dotnet/sdk/issues/26710#issuecomment-2676891024

It seems that the property used for this for web-based projects is typically set in a target, which means it isn't available at evaulation time. This changes the property used to determine if a project is web-based to one set by importing the web sdk.

Although I think it should not be used in almost any case, it also adds the option of specifying that the target should run after a specified target. If a new project type comes up that has an AfterAfterPublish, for instance, and we want to run this after that, this makes it easy to support that.