Suppress welcome message with get* builds Fixes #38444 (#47197)
Fixes #38444

This looks at the parseResult and scans it for any of the get* operators: getProperty, getItem, or getTargetResult. If it finds any of them (via string checks), it defers the first use message for the next operation. As the output from get* operators is often parsed, this dramatically simplifies that parsing.

The welcome message is still shown but does not appear in the part often parsed (the output of this get* command).