Check for Muxer path relative to base directory (#47664)
When determining the muxer (dotnet) path, search relative to the app base directory first. In most scenarios, this should correspond to the dotnet.dll directory of <root>/sdk/<version> where the dotnet executable is under <root>. If it doesn't exist, fall back to the existing logic.

With custom SDK search paths in global.json, the running dotnet process may not be the same as that corresponding to the running SDK. This updated logic finds the dotnet corresponding to the running SDK such that it can be properly propagated to anything launched via SDK commands (for example, build).