sln-add: Get project DefaultProjectTypeGuid (#47661)
Fixes #47638, #522

Description and impact
When adding a new project that is not known for the SDK or MSBuild, it should read the DefaultProjectTypeGuid property from the project file.

Regression
Yes - this worked before 9.0.2xx

Risk
Low – Currently it errors when adding unknown project types, unless the ProjectTypeGuid is specified on the project itself, or on the solution file.

Testing
Unit tests on this were ambiguous, so they were renamed to better match their cases.
I was also able to reproduce the bug...