Use pull_request_target event for GH Actions workflows
This removes the annoying "2 workflows awaiting approval".
We don't need to run these workflows with the "pull_request" trigger (which runs in the context of the fork and is therefore dangerous), the "pull_request_target" is enough (which runs in the context of the base branch and is safe).

We can also remove the source checkout from the remove-lockdown-label.yml since we don't read anything from the repo.