Respect BuildWithNetFrameworkHostedCompiler on netxf
This changes the build to only respect
BuildWithNetFrameworkHostedCompiler on full framework msbuild. Without
this guard customers can accidentally force the use of
Microsoft.Net.Sdk.Compilers.Toolset on a core msbuild. That leads to
runtime errors as the build tasks inside are .NET Framework specific.
These errors tend to be silent as they are in the server connection code
which will fall back to csc.exe invocation on error