Copilot refactor: remove extra method arguments
Refactor #method:'Microsoft.DotNet.Cli.ToolPackage.ToolPackageDownloaderBase.CreateNuGetPackageDownloader':2769-3057 and the calls to it to remove method parameters:

The packageInstallDir can always be a default DirectoryPath.  It doesn't need to differ based on whether a tool is global or local.

The creation of the ILogger can be moved into the implementation of the method.  Instead of passing an ILogger parameter, a verbosity parameter can be passed.