Fix API Diff args + async
APICompat was exiting at the first time it needed to go async.

This was because Commandline was treating the SetAction call as
synchronous since it didn't accept a CancellationToken.  Fix this and
plumb basic cancellation.

I also noticed that the usage of Option<T> was incorrect and corrected
it to be consistent with the way we use in other tools.