Add all of DNVM Src to the SDK
Ok, so we were going to make a bootstrapper that was standalone and could be invoked from the sdk. We started that in the cli-lab repo. Then budget cuts, layoffs, and office politics made it so we decided to put it into the SDK. Then we decided to use the existing code instead of taking the time to write all new code.

The point of this will be to have install and uninstall / local sdk management from the sdk itself for things like ci runners, making it easy to manage a local sdk or initialize a repo to download the sdk specified in a global json, etc.

All I am doing in this commit is copying everything from DNVM when it was MIT license as approved by the owner. That is step 1. If you review this all you need to check is that I copied the src folder correctly. Idk how you mess that up. Step 2 would be getting it to build. Step 3 will be taking out the parts we dont need such as serde. I am just copying it all so you can review it but none of this will be used by the SDK until later.