Create batch file for tools with runner=executable
Copilot prompt used to help generate batch file::

Can you write c# code that will generate a batch file?  Inputs should be the path to the batch file and the path to an executable that the batch file will run.  The C# code should find the relative path to the executable from the batch file path.  The batch file it generates should use that relative path to launch the executable and forward any arguments passed in to it.