Copilot: Create symbolic link for shims to executable tools
Copilot prompt:

Can you add code te the highlighted section of CreateShim to create a symlink from the shimPath to the toolCommand Executable?

Followup questions:

When using File.CreateSymbolicLink, should the pathToTarget parameter be an absolute or relative path?

When creating a symbolic link to an executable, do you need to set executable permission on the symbolic link?