Copilot: Set TargetFramework folder to "any" for SelfContained tools
Copilot prompt:

Rename _NuGetShortFolderName to _ToolPackShortTargetFrameworkName.  Update usages of the property.

Modify the _PackToolValidation target so that the _ToolPackShortTargetFrameworkName property is set to "any" if SelfContained is set to True.

Modify the #class:'Microsoft.NET.Build.Tasks.ResolveToolPackagePaths':583-4346 to take _ToolPackShortTargetFrameworkName as a parameter instead of the TargetFrameworkMoniker and TargetPlatformMoniker parameters.  Update the path construction to use the new parameter.  Update usage of the task to pass in the new parameter.