Copilot: Switch to AddPackageType task for adding DotnetTool package type
Copilot prompt:

Please add a task to #file:'Microsoft.NET.Build.Tasks.csproj' called AddPackageType.  It should take two parameters, CurrentPackageType which should be the current package type, and PackageTypeToAdd.  Both should be strings.

The logic in the task should be the same as what is currently in the SetDotnetToolPackageType target in #file:'Microsoft.NET.Sdk.Common.targets' .  It should check to see if the package type specified in PackageTypeToAdd is already in CurrentPackageType (whether or not it has a version).  If not, it should add the specified package type to the semicolon-separated list, otherwise it should leave it unchanged.  Eitherway the result should be returned inn an output parameter called UpdatedPackageType.

Then, update the logic in the SetDotnetToolPackageType to call the new task.  Don't forget to add a corresponding UsingTask declaration.